Add autorotate effect to custom image styles
We specifically do not want to modify a user's file on upload, because
that would be less than ideal from a digital preservation standpoint. We
do want to autorotate thumbnails and such.  There are quite a few images
being uploaded directly from phones that have the exif orientation field set.