Clear any duplicated book or CR relations
We don't present duplicate most of the time when these relations exist,
but some cases do exist. Previous behavior was to copy the relation,
not the full related node. This is problematic in the case of CRs,
because if the duplicated node is deleted, the CR will be deleted for
both the duplicated node and the original node.