Remove community record parent from DH importer
This field was too hastily added. There are two scenarios we are trying
to support. One in which a DH item is imported first and then
community records are imported second. Second where the CRs are imported
first, and the 'parent' DH node imported last. Having both 'Community
Record' and 'Community Record Parent' fields on the DH importer was an
attempt at supporting this, but having a double representation of the
relationship on a single CSV is too confusing for the user. We will add
a secondary importer for community record instead.