Make a specific set of local task tabs for each CR
Previously when viewing a DH item with CRs, only the parent's local task
tabs would appear. This change moves the local task tabs below the CR
selection interface and creates a CR specific set of tabs. When a user
selects a CR, the displayed tabs now correctly reflect the displayed CR.

This change depends on a new template variable 'cr_tabs'. An overloaded
template is included in the mukurtu_starter theme that utilizes this
variable. This change will need to be migrated for any custom themes.